+++ app/src/main/java/acr/browser/lightning/browser/tabs/TabViewState.kt
@@ -1,4 +1,4 @@
-package acr.browser.lightning.browser.fragment
+package acr.browser.lightning.browser.tabs
 
 import acr.browser.lightning.view.LightningView
 import android.graphics.Bitmap
