+++ app/src/main/java/acr/browser/lightning/browser/tabs/TabViewStateDiffCallback.kt
@@ -1,4 +1,4 @@
-package acr.browser.lightning.browser.fragment
+package acr.browser.lightning.browser.tabs
 
 import androidx.recyclerview.widget.DiffUtil
