+++ app/src/main/java/acr/browser/lightning/browser/tabs/TabsDrawerAdapter.kt
@@ -1,60 +1,34 @@
-package acr.browser.lightning.browser.fragment
+package acr.browser.lightning.browser.tabs
 
 import acr.browser.lightning.R
 import acr.browser.lightning.controller.UIController
 import acr.browser.lightning.extensions.desaturate
-import acr.browser.lightning.extensions.drawTrapezoid
 import acr.browser.lightning.extensions.inflater
-import acr.browser.lightning.utils.ThemeUtils
-import acr.browser.lightning.utils.Utils
-import android.content.Context
-import android.content.res.Resources
+import acr.browser.lightning.view.BackgroundDrawable
 import android.graphics.Bitmap
-import android.graphics.Canvas
-import android.graphics.Color
-import android.graphics.PorterDuff
-import android.graphics.drawable.BitmapDrawable
-import android.graphics.drawable.Drawable
 import android.view.ViewGroup
 import androidx.core.widget.TextViewCompat
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.RecyclerView
 
 /**
- * The adapter for horizontal desktop style browser tabs.
+ * The adapter for vertical mobile style browser tabs.
  */
-class TabsDesktopAdapter(
-    context: Context,
-    private val resources: Resources,
+class TabsDrawerAdapter(
     private val uiController: UIController
 ) : RecyclerView.Adapter<TabViewHolder>() {
 
-    private val backgroundTabDrawable: Drawable?
-    private val foregroundTabBitmap: Bitmap?
     private var tabList: List<TabViewState> = emptyList()
-
-    init {
-        val backgroundColor = Utils.mixTwoColors(ThemeUtils.getPrimaryColor(context), Color.BLACK, 0.75f)
-        val backgroundTabBitmap = Bitmap.createBitmap(Utils.dpToPx(175f), Utils.dpToPx(30f), Bitmap.Config.ARGB_8888).also {
-            Canvas(it).drawTrapezoid(backgroundColor, true)
-        }
-        backgroundTabDrawable = BitmapDrawable(resources, backgroundTabBitmap)
-
-        val foregroundColor = ThemeUtils.getPrimaryColor(context)
-        foregroundTabBitmap = Bitmap.createBitmap(Utils.dpToPx(175f), Utils.dpToPx(30f), Bitmap.Config.ARGB_8888).also {
-            Canvas(it).drawTrapezoid(foregroundColor, false)
-        }
-    }
 
     fun showTabs(tabs: List<TabViewState>) {
         val oldList = tabList
         tabList = tabs
-
         DiffUtil.calculateDiff(TabViewStateDiffCallback(oldList, tabList)).dispatchUpdatesTo(this)
     }
 
     override fun onCreateViewHolder(viewGroup: ViewGroup, i: Int): TabViewHolder {
-        val view = viewGroup.context.inflater.inflate(R.layout.tab_list_item_horizontal, viewGroup, false)
+        val view = viewGroup.context.inflater.inflate(R.layout.tab_list_item, viewGroup, false)
+        view.background = BackgroundDrawable(view.context)
         return TabViewHolder(view, uiController)
     }
 
@@ -66,6 +40,7 @@ class TabsDesktopAdapter(
         holder.txtTitle.text = web.title
         updateViewHolderAppearance(holder, web.favicon, web.isForegroundTab)
         updateViewHolderFavicon(holder, web.favicon, web.isForegroundTab)
+        updateViewHolderBackground(holder, web.isForegroundTab)
     }
 
     private fun updateViewHolderFavicon(viewHolder: TabViewHolder, favicon: Bitmap?, isForeground: Boolean) {
@@ -78,18 +53,22 @@ class TabsDesktopAdapter(
         } ?: viewHolder.favicon.setImageResource(R.drawable.ic_webpage)
     }
 
+    private fun updateViewHolderBackground(viewHolder: TabViewHolder, isForeground: Boolean) {
+        val verticalBackground = viewHolder.layout.background as BackgroundDrawable
+        verticalBackground.isCrossFadeEnabled = false
+        if (isForeground) {
+            verticalBackground.startTransition(200)
+        } else {
+            verticalBackground.reverseTransition(200)
+        }
+    }
+
     private fun updateViewHolderAppearance(viewHolder: TabViewHolder, favicon: Bitmap?, isForeground: Boolean) {
         if (isForeground) {
-            val foregroundDrawable = BitmapDrawable(resources, foregroundTabBitmap)
-            if (uiController.isColorMode()) {
-                foregroundDrawable.setColorFilter(uiController.getUiColor(), PorterDuff.Mode.SRC_IN)
-            }
             TextViewCompat.setTextAppearance(viewHolder.txtTitle, R.style.boldText)
-            viewHolder.layout.background = foregroundDrawable
-            uiController.changeToolbarBackground(favicon, foregroundDrawable)
+            uiController.changeToolbarBackground(favicon, null)
         } else {
             TextViewCompat.setTextAppearance(viewHolder.txtTitle, R.style.normalText)
-            viewHolder.layout.background = backgroundTabDrawable
         }
     }
